+++ src/account/TransferTransaction.js
@@ -13,7 +13,6 @@ import TokenNftTransfer from "../token/TokenNftTransfer.js";
 import NftId from "../token/NftId.js";
 import AbstractTokenTransferTransaction from "../token/AbstractTokenTransferTransaction.js";
 import FungibleHookCall from "../hooks/FungibleHookCall.js";
-import NftHookCall from "../hooks/NftHookCall.js";
 
 /**
  * @typedef {import("../long.js").LongObject} LongObject
@@ -32,7 +31,9 @@ import NftHookCall from "../hooks/NftHookCall.js";
 
 /**
  * @typedef {import("../channel/Channel.js").default} Channel
- * @typedef {import("../client/Client.js").default<*, *>} Client
+ * @typedef {import("../client/Client.js").default<Channel, import("../channel/MirrorChannel.js").default>} Client
+ * @typedef {import("../Timestamp.js").default} Timestamp
+ * @typedef {import("../hooks/NftHookCall.js").default} NftHookCall
  * @typedef {import("../transaction/TransactionId.js").default} TransactionId
  */
 
@@ -315,9 +316,8 @@ export default class TransferTransaction extends AbstractTokenTransferTransactio
     }
 
     /**
-     *
-     * @param {AccountId} accountId
-     * @param {Long} amount
+     * @param {AccountId | string} accountId
+     * @param {number | string | Long | LongObject | BigNumber | Hbar} amount
      * @param {FungibleHookCall} hook
      * @returns {TransferTransaction}
      */
@@ -336,10 +336,9 @@ export default class TransferTransaction extends AbstractTokenTransferTransactio
     }
 
     /**
-     *
-     * @param {NftId} nftId
-     * @param {AccountId} sender
-     * @param {AccountId} receiver
+     * @param {NftId | string} nftId
+     * @param {AccountId | string} sender
+     * @param {AccountId | string} receiver
      * @param {NftHookCall} senderHookCall
      * @param {NftHookCall} receiverHookCall
      * @returns {TransferTransaction}
@@ -351,11 +350,21 @@ export default class TransferTransaction extends AbstractTokenTransferTransactio
         senderHookCall,
         receiverHookCall,
     ) {
+        const nft = nftId instanceof NftId ? nftId : NftId.fromString(nftId);
+        const senderId =
+            sender instanceof AccountId
+                ? sender.clone()
+                : AccountId.fromString(sender);
+        const receiverId =
+            receiver instanceof AccountId
+                ? receiver.clone()
+                : AccountId.fromString(receiver);
+
         return this._addNftTransfer(
             false,
-            nftId,
-            sender,
-            receiver,
+            nft,
+            senderId,
+            receiverId,
             undefined, // receiver
             senderHookCall,
             receiverHookCall,
@@ -363,14 +372,22 @@ export default class TransferTransaction extends AbstractTokenTransferTransactio
     }
 
     /**
-     *
-     * @param {TokenId} tokenId
-     * @param {AccountId} accountId
+     * @param {TokenId | string} tokenId
+     * @param {AccountId | string} accountId
      * @param {Long} amount
      * @param {FungibleHookCall} hook
      * @returns {TransferTransaction}
      */
     addTokenTransferWithHook(tokenId, accountId, amount, hook) {
+        const token =
+            tokenId instanceof TokenId
+                ? tokenId.clone()
+                : TokenId.fromString(tokenId);
+        const account =
+            accountId instanceof AccountId
+                ? accountId.clone()
+                : AccountId.fromString(accountId);
+
         const fungibleHook = new FungibleHookCall({
             hookId: hook.hookId != null ? hook.hookId : undefined,
             evmHookCall:
@@ -379,17 +396,18 @@ export default class TransferTransaction extends AbstractTokenTransferTransactio
         });
 
         const isApproved = false; // this is not approved transfer, adding comment for clarity
-        const expectedDecimals = null; // we don't expect  decimals here, adding comment for clarity
+        const expectedDecimals = null; // we don't expect decimals here, adding comment for clarity
 
         return this._addTokenTransfer(
-            tokenId,
-            accountId,
+            token,
+            account,
             amount,
             isApproved,
             expectedDecimals,
             fungibleHook,
         );
     }
+
     /**
      * @override
      * @internal
@@ -431,10 +449,11 @@ export default class TransferTransaction extends AbstractTokenTransferTransactio
     }
 
     /**
+     * @override
      * @returns {string}
      */
     _getLogId() {
-        const timestamp = /** @type {import("../Timestamp.js").default} */ (
+        const timestamp = /** @type {Timestamp} */ (
             this._transactionIds.current.validStart
         );
         return `TransferTransaction:${timestamp.toString()}`;
@@ -443,6 +462,5 @@ export default class TransferTransaction extends AbstractTokenTransferTransactio
 
 TRANSACTION_REGISTRY.set(
     "cryptoTransfer",
-    // eslint-disable-next-line @typescript-eslint/unbound-method
-    TransferTransaction._fromProtobuf,
+    TransferTransaction._fromProtobuf.bind(TransferTransaction),
 );